NEWS	R Documentation

News for package ‘panelPomp’

Changes in panelPomp version 1.1.0.2

Added a type check for the params arguement of the pfilter function in order to throw more user friendly error.

A more user friendly error is thrown when the user specifies a parameter as both shared and unit specific in the mif2 function.

Added a match.arg call in mif2.internal; this will allow the default option for the cooling.type argument in mif2 to no longer throw an error, and mimics the behaviour of the mif2 function in the pomp package.

The pParams function no longer changes the order of the unit specific parameter names; this was causing an error because the barycentric parameter transformation requires that the parameters are adjacent to one another in the parameter vector.

Improved documentation for panel_logmeanexp.

More robust input checking for the pfilter function with user friendly error messages.

Changes in panelPomp version 1.1.0

Improved documentation and tests.

Changes in panelPomp version 1.0.0

Submitted to CRAN.

Changes in panelPomp version 0.18.0

Fixed bug in fromVectorPparams() to make it a true inverse of toVectorPparams().

Homogenized documentation; added examples.

Changes in panelPomp version 0.15.0

New coercion to pompList and update to list.

panelGompertzLikelihood replaces kalman_gompertz.R.

expunged panelPompExample.

Names added to plot; pfilter matches units by name.

Changes in panelPomp version 0.14.0

New simulate function.

Changes in panelPomp version 0.13.0

New panelRandomWalk example.

panelPompExample is now deprecated and will be removed in the near future.

Changes in panelPomp version 0.12.0

New plot method for panelPomp objects.

Changes in panelPomp version 0.11.0

Additional functions implementing Kalman filtering to panels and to the Gompertz model.

Changes in panelPomp version 0.10.2.0

Minor improvements to the interface.

Additional examples: contacts() and panelGompertz().

The minimum version of pomp supported is now 2.7.

Changes in panelPomp version 0.9.1

The minimum version of pomp supported is now 2.1. While this is likely to imply substantial changes in pomp objects (on which panelPomp objects rely) there are only three fundamental changes that break panelPomp's backwards compatibility:

The first argument pfilter(object, ...) changes to data.

The first argument mif2(object, ...) changes to data.

The ‘conv.rec’ method has been deprecated, replaced by the new ‘traces’ method.

Changes in panelPomp version 0.8.3

Improvements to the interface.

Changes in panelPomp version 0.7

A new conv.rec method.

Changes in panelPomp version 0.6

Many improvements to the interface.

Changes in panelPomp version 0.5.6

panelPompExamples() now allows creating panelPomp objects that come with the package.

New documentation and tests.

Changes in panelPomp version 0.5.5

New documentation and tests.

Changes in panelPomp version 0.5.4

New documentation and tests.

Changes in panelPomp version 0.5.3

New documentation and additional tests.

Changes in panelPomp version 0.5

New documentation and additional tests and coverage.

Changes in panelPomp version 0.4

New tests added.

New methods.

A bug in pfilter and mif2 has been fixed (pomp's default filtering tolerance tol=1e-17 was used instead of values specified by the user). Thanks to Aaron King for finding it.

Changes in panelPomp version 0.2

New tests added.

Eliminate need to protect rw.sd from being evaluated.

Changes in panelPomp version 0.1

Unused variable purged.

pPfilter.internal streamlined.

Test for reproducibility of pfilter calculations added.