as {panelPomp}	R Documentation

Coercing panelPomp objects as a list

Description

Extracts the unit.objects slot of panelPomp objects and attaches the associated parameters.

Extracts the unit.objects slot of panelPomp objects and attaches the associated parameters, converting the resulting list to a pompList to help the assignment of pomp methods.

Coerces a panelPomp into a data frame, assuming units share common variable names.

See Also

Other panelPomp methods: panelPomp_methods

Other panelPomp methods: panelPomp_methods

Other panelPomp methods: panelPomp_methods

[Package panelPomp version 0.15.0 Index]