get_dim {panelPomp}	R Documentation

Get single column or row without dropping names

Description

Subset matrix dropping dimension but without dropping dimname (which is R's default).

Usage

get_col(matrix, rows, col)

get_row(matrix, row, cols)

Arguments

matrix
matrix.

rows
numeric; rows to subset; like with '[', this argument can be left empty to designate all rows.

col
numeric; single column to subset.

row
numeric; single row to subset.

cols
numeric; columns to subset; like with '[', this argument can be left empty to designate all columns.

[Package panelPomp version 0.15.0 Index]