panel-designs {panelPomp}	R Documentation

#' Create design matrix for panelPomp calculations

Description

These functions are useful for generating design matrices for the exploration of parameter space.

Usage

runif_panel_design(
  lower = numeric(0),
  upper = numeric(0),
  nseq,
  specific_names,
  unit_names
)

Arguments

lower, upper
named numeric vectors giving the lower and upper bounds of the ranges, respectively.

nseq
Total number of points requested

specific_names
Character vector containing the names of unit-specific parameters. This argument must be used in conjunction with the argument unit_names; it is used if the search bounds for all unspecified unit-specific parameters are the same.

unit_names
Character vector containing the names of the units of the panel. If not used in conjunction with unit_names this argument is ignored.

Value

runif_panel_design returns a data.frame object with nseq rows. Each row corresponds to a parameter set drawn randomly from a multivariate uniform distribution specified by the lower, upper, specific_names and unit_names arguments.

Author(s)

Jesse Wheeler, Aaron A. King

Examples

runif_panel_design(
  lower = c('a' = 0, 'b' = 10, 'a[u2]' = 0.5),
  upper = c('a' = 1, 'b' = 15, 'a[u2]' = 0.75),
  specific_names = c('a'),
  unit_names = paste0(rep('u', 5), 1:5),
  nseq = 10
)

[Package panelPomp version 1.2.0 Index]