panelGompertz {panelPomp}	R Documentation

Panel Gompertz model

Description

Builds a collection of independent realizations from the Gompertz model.

Usage

panelGompertz(
  N = 100,
  U = 50,
  params = c(K = 1, r = 0.1, sigma = 0.1, tau = 0.1, X.0 = 1),
  seed = 12345678
)

Arguments

N
number of observations for each unit.

U
number of units.

params
parameter vector, assuming all units have the same parameters.

seed
passed to the random number generator for simulation.

Value

A panelPomp object.

Author(s)

Edward L. Ionides, Carles Bretó

References

Bretó, C., Ionides, E. L. and King, A. A. (2020) Panel Data Analysis via Mechanistic Models. Journal of the American Statistical Association, 115(531), 1178–1188. doi:10.1080/01621459.2019.1604367

King, A. A., Nguyen, D. and Ionides, E. L. (2016) Statistical inference for partially observed Markov processes via the package pomp. Journal of Statistical Software 69(12), 1–43. DOI: 10.18637/jss.v069.i12. An updated version of this paper is available on the package website.

See Also

Other panelPomp examples: contacts(), panelRandomWalk()

Examples

panelGompertz()

[Package panelPomp version 1.2.0 Index]