panelGompertzLikelihood {panelPomp}	R Documentation

Likelihood for a panel Gompertz model via a Kalman filter

Description

Evaluates the likelihood function for a panel Gompertz model, using a format convenient for maximization by optim() to obtain a maximum likelihood estimate. Specifically, estimated and fixed parameters are supplied by two different arguments.

Usage

panelGompertzLikelihood(x, panelPompObject, params)

Arguments

x
named vector for a subset of parameters, corresponding to those being estimated.

panelPompObject
a panel Gompertz model.

params
named vector containing all the parameters of the panel Gompertz model. Estimated parameters are overwritten by x.

Value

A numeric value.

Author(s)

Edward L. Ionides

Examples

pg <- panelGompertz(N=2,U=2)
panelGompertzLikelihood(coef(pg),pg,coef(pg))

[Package panelPomp version 1.2.0 Index]