panelMeasles {panelPomp}	R Documentation

Make a panelPomp model using UK measles data.

Description

The model is a modified panelPomp version of the model of He et al. 2010. The model is a stochastic SEIR model that accounts for population demographics in the form of births and deaths. Because of the increased transmission that results from school-aged children entering the susceptible pool once they begin attending classes for the first time, the model includes a birth-cohort effect, which moves a specified faction of the cohort into the susceptible pool all at once. The model also includes a seasonality in transmission rate that is larger during school terms thn it is during holidays.

Usage

panelMeasles(
  units = c("Bedwellty", "Birmingham", "Bradford", "Bristol", "Cardiff", "Consett",
    "Dalton.in.Furness", "Halesworth", "Hastings", "Hull", "Leeds", "Lees", "Liverpool",
    "London", "Manchester", "Mold", "Northwich", "Nottingham", "Oswestry", "Sheffield"),
  starting_pparams = NULL,
  interp_method = c("shifted_splines", "linear"),
  first_year = 1950,
  last_year = 1963,
  dt = 1/365.25
)

Arguments

units
Character vector of units in uk_measles to be used in the panel model.

starting_pparams
Parameters in the list format, having shared and specific components. Set to NULL to assign NA values.

interp_method
Method used to interpolate population and births. Possible options are "shifted_splines" and "linear".

first_year
Integer for the first full year of data desired.

last_year
Integer for the last full year of data desired.

dt
Size of the time step.

Value

A panelPomp object.

References

D. He, E.L. Ionides, and A.A. King. Plug-and-play inference for disease dynamics: measles in large and small populations as a case study. Journal of the Royal Society Interface 7, 271–283, 2010.

See Also

Other panelPomp examples: contacts(), panelGompertz(), panelRandomWalk()

Examples

panelMeasles(units = "London")

[Package panelPomp version 1.6.0.0 Index]