panelRandomWalk {panelPomp}	R Documentation

Panel random walk model

Description

Builds a collection of independent realizations from a random walk model.

Usage

panelRandomWalk(
  N = 5,
  U = 2,
  params = c(sigmaY = 1, sigmaX = 1, X.0 = 1),
  seed = 3141592
)

Arguments

N
number of observations for each unit.

U
number of units.

params
parameter vector, assuming all units have the same parameters.

seed
passed to the random number generator for simulation.

Value

A panelPomp object.

Author(s)

Edward L. Ionides, Carles Bretó

See Also

Other panelPomp examples: contacts(), panelGompertz()

Examples

panelRandomWalk()

[Package panelPomp version 1.1.0.2 Index]