panel_loglik {panelPomp}	R Documentation

Handling of loglikelihood replicates

Description

Handling of loglikelihood replicates.

Usage

## S4 method for signature 'matrix'
logLik(object, repMargin, first = "aver", aver = "logmeanexp", se = FALSE)

Arguments

object
Matrix with the same number of replicated estimates for each panel unit loglikelihood.

repMargin
The margin of the matrix having the replicates (1 for rows, 2 for columns).

first
Wether to "aver"(age replicates) or "aggr"(egate units) before performing the other action.

aver
How to average: 'logmeanexp' to average on the likelihood scale before taking logs or 'mean' to average after taking logs (in which case, which action is performed first does not change the result).

se
logical; whether to give standard errors.

Details

When se = TRUE, the jackknife se's from pomp::logmeanexp are squared, summed and the squared root is taken.

See Also

Other panelPomp workhorse functions: mif2(), panelPomp, pfilter()

[Package panelPomp version 0.15.0 Index]