params {panelPomp}	R Documentation

Convert to and from a panelPomp object pParams slot format and a one-row data.frame

Description

These facilitate keeping a record of evaluated log likelihoods.

Usage

fromVectorPparams(vec_pars)

toMatrixPparams(listPparams)

toVectorPparams(pParams)

Arguments

vec_pars
A one-row data.frame with format matching that of the output of toVectorPparams.

listPparams
PanelPomp parameters in list format

pParams
A list with the format of the pParams slot of panelPomp objects.

[Package panelPomp version 0.16.0 Index]