coef<-,pfilterd.ppomp-method {panelPomp}	R Documentation

Modifying parameters of filtered objects

Description

The setter functions for parameters of pfilterd.ppomp objects do not allow users to set parameters of panelPomp objects that have been filtered. This is done to avoid the possibility of having parameter values in an object that do not match other attributes of a filtered object to be saved together.

The setter functions for parameters of pfilterd.ppomp objects do not allow users to set parameters of panelPomp objects that have been filtered. This is done to avoid the possibility of having parameter values in an object that do not match other attributes of a filtered object to be saved together.

The setter functions for parameters of pfilterd.ppomp objects do not allow users to set parameters of panelPomp objects that have been filtered. This is done to avoid the possibility of having parameter values in an object that do not match other attributes of a filtered object to be saved together.

Usage

## S4 replacement method for signature 'pfilterd.ppomp'
coef(object, ...) <- value

## S4 replacement method for signature 'pfilterd.ppomp'
shared(object) <- value

## S4 replacement method for signature 'pfilterd.ppomp'
specific(object) <- value

Arguments

object
pfilterd.ppomp object

...
additional arguments.

value
New parameter value. This function does not allow users to set this value.

[Package panelPomp version 1.2.0 Index]