pparams {panelPomp}	R Documentation

Extract parameters (coefficients) of a panel model

Description

pparams() is a generic function that extracts parameter (coefficient) values from objects returned by panel modeling functions. While the named numeric vector format is useful and possible via S4 methods for coef(), alternative formats capturing the panel structure can be implemented via pparams().

Usage

pparams(object, ...)

Arguments

object
an object for which extraction of panel model parameter (coefficient) values is meaningful.

...
additional arguments.

Details

This is a generic function: methods can be defined for it.

Value

Parameter (coefficient) values extracted from the panel model object.

When given objects of class panelPomp, pparams() returns a named list with two elements: shared, which is a named numeric vector, and specific, which is a matrix with parameter names in its row names and panel unit names in its column names.

Author(s)

Carles Bretó

See Also

panelPomp_methods

Examples

prw <- panelRandomWalk()
# extract parameters in list form
pparams(prw)

[Package panelPomp version 1.2.0 Index]