shared<- {panelPomp}	R Documentation

Set shared parameters of a panelPomp object

Description

This function is used to set the shared parameters of a panel pomp object.

Usage

shared(object) <- value

Arguments

object
an object that contains shared parameters.

value
a named numeric vector containing the desired values of the shared parameter vector.

Author(s)

Jesse Wheeler

See Also

panelPomp_methods

Examples

prw <- panelRandomWalk()
# set parameters in list form
shared(prw) <- c("sigmaX" = 1, "sigmaY" = 2)

[Package panelPomp version 1.2.0 Index]