simulate {panelPomp}	R Documentation

Simulations of a panel of partially observed Markov process

Description

simulate generates simulations of the state and measurement processes.

Usage

## S4 method for signature 'panelPomp'
simulate(object, nsim = 1, shared, specific)

Arguments

object
a panelPomp object.

nsim
The number of simulations to perform. Unlike the pomp simulate method, all simulations share the same parameters.

shared
Named vector of the shared parameters.

specific
Matrix of unit-specific parameters, with a column for each unit.

Value

A single panelPomp object (if nsim=1) or a list of panelPomp objects (if nsim>1).

Author(s)

Edward L. Ionides

Examples

simulate(panelRandomWalk())

[Package panelPomp version 1.1.0.2 Index]