specific<- {panelPomp}	R Documentation

Set unit-specific parameters of a panelPomp object

Description

This function is used to set the unit-specific parameters of a panel pomp object.

Usage

specific(object) <- value

Arguments

object
an object that contains unit-specific parameters.

value
a numeric matrix with column names matching the names of the unit_objects slot, and row names matching the names of the unit-specific parameters. Alternatively, this can be a named vector following the naming convention ⁠<parameter>[<unit_name>]⁠.

Author(s)

Jesse Wheeler

See Also

panelPomp_methods

Examples

# set parameters in list form
prw <- panelRandomWalk(U = 4)
new_pars <- matrix(c(1, 1, 3, 2), nrow = 1)
dimnames(new_pars) <- list(param = "X.0", unit = c("rw1", "rw2", "rw3", "rw4"))
specific(prw) <- new_pars

[Package panelPomp version 1.2.0 Index]