unitlogLik-deprecated {panelPomp}	R Documentation

(Deprecated) Extract log likelihood of units of panel models

Description

In future versions, this generic is being replaced with unitLogLik().

unitlogLik() is a generic function that extracts the log likelihood for each unit of panel objects returned by panel modeling functions. While the numeric value with the log likelihood for the entire panel is useful and possible via S4 methods logLik(), the contributions to it by panel units can be implemented via unitlogLik().

Usage

unitlogLik(object, ...)

Arguments

object
an object for which log likelihood values for units can be extracted.

...
additional arguments.

Details

This is a generic function: methods can be defined for it.

Value

Log likelihood extracted for each unit of the panel model object.

When given objects of class pfilterd.ppomp, unitloglik() returns a numeric vector.

Author(s)

Carles Bretó

See Also

pfilter

Examples

# filter, which generates log likelihoods
pfrw <- pfilter(panelRandomWalk(),Np=10)

# extract log likelihood for each panel unit
unitlogLik(pfrw)

[Package panelPomp version 1.2.0 Index]