unitobjects {panelPomp}	R Documentation

Extract units of a panel model

Description

unitobjects() is a generic function that extracts a list of objects corresponding to units of panel objects returned by panel modeling functions.

Usage

unitobjects(object, ...)

Arguments

object
an object for which extraction of panel units is meaningful.

...
additional arguments.

Details

This is a generic function: methods can be defined for it.

Value

Units extracted from the panel model object.

When given objects of class panelPomp, unitobjects() returns a list of pomp objects.

Author(s)

Carles Bretó

See Also

panelPomp_methods

Examples

prw <- panelRandomWalk()
## access underlying pomp objects
unitobjects(prw)

[Package panelPomp version 1.1.0.2 Index]