as {panelPomp}	R Documentation

Coercing panelPomp objects as list, pompList or data.frame

Description

When coercing to a data.frame, it coerces a panelPomp into a data.frame, assuming units share common variable names.

When coercing to a list, it extracts the unit_objects slot of panelPomp objects and attaches associated parameters.

When coercing to a pompList, it extracts the unit_objects slot of panelPomp objects and attaches associated parameters, converting the resulting list to a pompList to help the assignment of pomp methods.

Value

An object of class matching that specified in the second argument (to=).

Author(s)

Carles Bretó

See Also

Other panelPomp methods: panelPomp_methods

[Package panelPomp version 1.2.0 Index]