panel_logmeanexp {panelPomp}	R Documentation

Log-mean-exp for panels

Description

This function computes the logmeanexp for each column or row of a numeric matrix and sums the result. Because the loglikelihood of a panelPomp object is the sum of the loglikelihoods of its units, this function can be used to summarize replicated estimates of the panelPomp model likelihood. If se = TRUE, the jackknife SE estimates from logmeanexp are squared, summed and the squared root is taken.

Usage

panel_logmeanexp(x, MARGIN, se = FALSE)

Arguments

x
Matrix with the same number of replicated estimates for each panel unit loglikelihood.

MARGIN
The dimension of the matrix that corresponds to a panel unit and over which averaging occurs (1 indicates rows, 2 indicates columns).

se
logical; whether to give standard errors.

Value

A numeric value with the average panel log likelihood or, when se = TRUE, a numeric vector adding the corresponding standard error.

Author(s)

Carles Bretó

See Also

panel_loglik

Examples

ulls <- matrix(c(1,1,10,10),nr=2)
panel_logmeanexp(ulls,MARGIN=2,se=TRUE)

[Package panelPomp version 1.1.0.2 Index]