panel_logmeanexp {panelPomp}	R Documentation

Log-mean-exp for panels

Description

se = TRUE, the jackknife se's from logmeanexp are squared, summed and the squared root is taken.

Usage

panel_logmeanexp(x, MARGIN, se = FALSE)

Arguments

x
Matrix with the same number of replicated estimates for each panel unit loglikelihood.

MARGIN
The dimension of the matrix that corresponds to a panel unit and over which averaging occurs (1 indicates rows, 2 indicates columns).

se
logical; whether to give standard errors.

See Also

panel_loglik

[Package panelPomp version 0.15.1 Index]