params {panelPomp}	R Documentation

Convert to and from a panelPomp object pParams slot format and a one-row data.frame

Description

These facilitate keeping a record of evaluated log likelihoods.

Usage

toVectorPparams(pParams)

fromVectorPparams(vec_pars)

toMatrixPparams(listPparams)

Arguments

pParams
A list with the format of the pParams slot of panelPomp objects.

vec_pars
A one-row data.frame with format matching that of the output of toVectorPparams.

listPparams
PanelPomp parameters in list format

Value

toVectorPparams() returns an object of class data.frame.

fromVectorPparams() returns an object of class list with the model parameters in list form.

toMatrixPparams() returns an object of class matrix with the model parameters in matrix form.

Author(s)

Carles Bretó

Examples

prw <- panelRandomWalk()
toVectorPparams(pparams(prw))
fromVectorPparams(toVectorPparams(pparams(prw)))
toMatrixPparams(pparams(prw))

[Package panelPomp version 1.1.0.2 Index]