params {panelPomp}	R Documentation

Manipulating panelPomp object parameter formats

Description

These facilitate keeping a record of evaluated log likelihoods.

Usage

toParamVec(pParams)

toMatrixPparams(listPparams)

Arguments

pParams
A list with both shared (vector) and unit-specific (matrix) parameters.

listPparams
PanelPomp parameters in list format

Value

toParamVec() returns model parameters in vector form. This function is the inverse of toParamList

toMatrixPparams() returns an object of class matrix with the model parameters in matrix form.

Author(s)

Carles Bretó

Examples

prw <- panelRandomWalk()
toParamVec(coef(prw, format = 'list'))
toMatrixPparams(coef(prw, format = 'list'))

[Package panelPomp version 1.6.0.0 Index]